Port RestorablePackageSet to packages module
RestorablePackageSet is moved to be under the packages module inside the
backup folder. This is one part of the refactoring of the backup
controller to make it more organized and easier to work with.
